// src/siteapi/badges.v1/models/RobloxWebWebAPIModelsApiArrayResponseRobloxBadgesApiBadgeAwardResponse.ts
/**
 * Badges Api v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RobloxBadgesApiBadgeAwardResponse } from '../models/RobloxBadgesApiBadgeAwardResponse';
import { HttpFile } from '../../../common/http/http';

export class RobloxWebWebAPIModelsApiArrayResponseRobloxBadgesApiBadgeAwardResponse {
    'data'?: Array<RobloxBadgesApiBadgeAwardResponse>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<RobloxBadgesApiBadgeAwardResponse>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RobloxWebWebAPIModelsApiArrayResponseRobloxBadgesApiBadgeAwardResponse.attributeTypeMap;
    }

    public constructor() {
    }
}