// src/siteapi/badges.v1/models/RobloxWebWebAPIModelsApiPageResponseRobloxBadgesApiBadgeResponse.ts
/**
 * Badges Api v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RobloxBadgesApiBadgeResponse } from '../models/RobloxBadgesApiBadgeResponse';
import { HttpFile } from '../../../common/http/http';

export class RobloxWebWebAPIModelsApiPageResponseRobloxBadgesApiBadgeResponse {
    'previousPageCursor'?: string;
    'nextPageCursor'?: string;
    'data'?: Array<RobloxBadgesApiBadgeResponse>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "previousPageCursor",
            "baseName": "previousPageCursor",
            "type": "string",
            "format": ""
        },
        {
            "name": "nextPageCursor",
            "baseName": "nextPageCursor",
            "type": "string",
            "format": ""
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<RobloxBadgesApiBadgeResponse>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RobloxWebWebAPIModelsApiPageResponseRobloxBadgesApiBadgeResponse.attributeTypeMap;
    }

    public constructor() {
    }
}